+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Cutscene max tick conflict if removing an animation/subscene
 - Revert fix on story (break choices appear)
 - World locked if leaving the world while a playback is playing
+- Character placement on ground in camera angle not accurate